% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_chromat.R
\name{geom_chromat}
\alias{geom_chromat}
\title{Display Chromatographs from LC-MS}
\usage{
geom_chromat(
  data,
  legend.position = "none",
  max.over = 30,
  guide.ncol = 1,
  xlimit = NA,
  x.nbreaks = 5,
  label = F,
  batchlabel = T,
  grid = F,
  shift = F,
  angle = 0,
  percent = 0.05,
  integration = F
)
}
\arguments{
\item{data}{data frame. Contain X(min), Y(Counts), batch and Compound.}

\item{legend.position}{the default position of legends ("none", "left", "right", "bottom", "top", "inside"). A numeric vector of length two setting the placement of legends that have the "inside" position.}

\item{max.over}{for group label. 30 by default. Only available when number of group is smaller than 30.}

\item{guide.ncol}{column number of legend.}

\item{xlimit}{limits for x axis}

\item{x.nbreaks}{An integer guiding the number of major breaks for x axis. 5 by default.}

\item{label}{label for group.}

\item{batchlabel}{label for batch.}

\item{grid}{logical. Should grid be displayed?}

\item{shift}{logical. Shift chromatographs from different batches. Available when batch more than 1.}

\item{angle}{angle of shift.}

\item{percent}{minor degree for shift}

\item{integration}{logical. Should area of chromatographs be displayed?}
}
\value{
a ggplot element
}
\description{
Display Chromatographs from LC-MS
}
\examples{
tmp <- chromatograph \%>\% filter(batch \%in\% c("HB-1","HB-2"))
geom_chromat(tmp, shift = F, angle = 45)

geom_chromat(tmp, shift = T, angle = 45, percent = 0.1)
}
