% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_MS2.R
\name{geom_MS2}
\alias{geom_MS2}
\title{Mass Spectrometry Dissociation Graphic}
\usage{
geom_MS2(
  dat,
  img = "",
  img.size = 30,
  heights = c(1, 1),
  colors = c("#222222", "#C4121A")
)
}
\arguments{
\item{dat}{a data frame, containing 4 columns named Compound,    m/z, Relative Intensity, Color.}

\item{img}{a character refers to the path of a picture of dissociation.}

\item{img.size}{size of dissociation picture.}

\item{heights}{a vector refers to heights of picture and dissociation graphic, respectively.}

\item{colors}{a vector used to customize color for refer and selected.}
}
\value{
a ggplot element
}
\description{
This function draws plots of mass spectrometry dissociation based on ggplot2 system.
}
\examples{
geom_MS2(MS2_sample)
}
