% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_MS2_2.R
\name{geom_MS2_2}
\alias{geom_MS2_2}
\title{Compare two Mass Spectrometry Dissociation Graphic}
\usage{
geom_MS2_2(
  dat,
  img = "",
  img.size = 30,
  heights = c(1, 1),
  cutoff = 0,
  colors = c("#222222", "#C4121A", "#006eb3", "#C4121A")
)
}
\arguments{
\item{img}{a character refers to the path of a picture of dissociation.}

\item{img.size}{size of dissociation picture.}

\item{heights}{a vector refers to heights of picture and dissociation graphic, respectively.}

\item{cutoff}{a threshold to filter fragment whose relative intensity under it}

\item{colors}{a vector used to customize color for sample-refer, sample-selected, refer-refer and refer-selected.}

\item{data}{data frame, containing 5 columns named Compound,    m/z, Relative Intensity, Color and group. Color is grouped by "refer" and "selected", and group is grouped by "sample" and "refer".}
}
\value{
a ggplot element
}
\description{
Compare two Mass Spectrometry Dissociation Graphic
}
\examples{
geom_MS2_2(MS2_compar, cutoff = 20)
}
