% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_multi.R
\name{intersect_multi}
\alias{intersect_multi}
\title{Intersect multiple elements}
\usage{
intersect_multi(list)
}
\arguments{
\item{list}{a list of elements to intersect}
}
\value{
in terms of input return vector or data.frame
}
\description{
Intersect multiple elements
}
\examples{
intersect_multi(list = list(c(1:3),c(1:4),c(2:4)))

intersect_multi(list = list(disso_sample, disso_sample, disso_sample))

}
