% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_chromat.R
\name{stat_chromat}
\alias{stat_chromat}
\title{Compute similarity of two chromatographs}
\usage{
stat_chromat(a, b, Na, Nb, Nab, method = "pearson")
}
\arguments{
\item{a}{Area or points of Chromatograph 1}

\item{b}{Area or points of Chromatograph 2}

\item{Na}{Number of peaks in Chromatograph 1}

\item{Nb}{Number of peaks in Chromatograph 2}

\item{Nab}{Number of peaks in both Chromatograph 1 and 2}

\item{method}{pearson, spearman, cosine, Overlap and Tanimoto methods are currently available}
}
\value{
a numeric value.
}
\description{
To compute similarity of two chromatographs
}
\examples{
stat_chromat(Na = 10, Nb = 11, Nab = 9, method = "Tanimoto")
}
