% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_disso2.R
\name{stat_disso2}
\alias{stat_disso2}
\title{Compute Cosine Similarity of two Mass Spectrometry Dissociation Graphic}
\usage{
stat_disso2(ref, sample, ref.cutoff = 10, ppm = 10)
}
\arguments{
\item{ref}{data frame of reference MS dissocaiation.}

\item{sample}{data frame of sample MS dissocaiation.}

\item{ref.cutoff}{Remove fragments whose relative intensity under cut-off.}

\item{ppm}{To screen the same fragment of reference in sample, 10 by default.}
}
\value{
a number between -1 and 1.
}
\description{
Compute Cosine Similarity of two Mass Spectrometry Dissociation Graphic
}
\examples{
stat_disso2(disso_reference, disso_sample, ref.cutoff = 30, ppm = 10000)
}
