% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_mol_weight.R
\name{stat_mol_weight}
\alias{stat_mol_weight}
\title{Compute molecular weight}
\usage{
stat_mol_weight(..., atom_mass = annot, adduct = F)
}
\arguments{
\item{...}{a vector or list of molecular formulas in character format}

\item{atom_mass}{a data frame includes exact atomic mass. Average Molecular Weight derived from https://iupac.qmul.ac.uk/AtWt/ was used by default}

\item{adduct}{If TRUE, compute molecular weight of adduct ions}
}
\value{
a data.frame includes peaks of sample matched in reference library
}
\description{
Compute molecular weight
}
\examples{
stat_mol_weight(c("CH3COOH","NaCl","H2O","2H2O","90745","/"))

stat_mol_weight(list(a = "CH3COOH", b = c("CH3CH2OH","KOH","Mg2+","[CH3COO]-")))

stat_mol_weight("CH3Cl","CH3CH2COOH", adduct = T)
}
