% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_MS1.R
\name{stat_MS1}
\alias{stat_MS1}
\title{Compare MS1 vs Reference Library}
\usage{
stat_MS1(
  REF = "reference_path",
  Sample = "sample_path",
  Filter_Adduct = c("M+H", "M-2H2O+H", "M-H2O+H", "M+Li", "M+NH4", "M+Na"),
  Filter_Charge = c("1+", "2+", "3+"),
  error = 0.1,
  use_ppm = F,
  atom_mass = annot
)
}
\arguments{
\item{REF}{directory of reference library file}

\item{Sample}{directory of sample file}

\item{Filter_Adduct}{the types of adduct ions should be considered. if "all", a total of 52 adduct ions will be computed}

\item{Filter_Charge}{the charges of adduct ions should be considered. if "all", all of adduct ions will be computed}

\item{error}{accepted error between reference and sample. Da by default.}

\item{use_ppm}{logical. if FALSE, the unit of error is Da, if TRUE, the unit of error is ppm.}

\item{atom_mass}{a data frame includes exact atomic mass. Average Molecular Weight derived from https://iupac.qmul.ac.uk/AtWt/ was used by default}
}
\value{
a list
}
\description{
This function compares MS1 derived from R package xcms vs reference library.
}
\examples{
no examples
}
