% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_mol_weight.R
\name{stat_mol_weight}
\alias{stat_mol_weight}
\title{Compute molecular weight}
\usage{
stat_mol_weight(...)
}
\arguments{
\item{...}{a vector or list of molecular formulas in character format}
}
\value{
list
}
\description{
Compute molecular weight
}
\examples{
stat_mol_weight(c("CH3COOH","NaCl","H2O","2H2O","90745","/"))

stat_mol_weight(list(a = "CH3COOH", b = c("CH3CH2OH","KOH","Mg2+","[CH3COO]-")))
}
